can cut image data by rows and columns also keeps track of best yet and writes output files

+++ model.r
@@ -2,6 +2,7 @@ library(plyr)
 library(nnet)
 library(caret)
 library(R.utils)
+library(foreach)
 
 set.seed(12345)
 
@@ -9,19 +10,42 @@ set.seed(12345)
 all.rows <- 28
 all.cols <- 28
 
-every <- 1
+every <- 4
 rows <- floor(all.rows / every)
 cols <- floor(all.cols / every)
 
-if (!exists("train.data")) {
-    max.rows <- 10000
+featureEngineering <- function (df) {
+    if (every > 1) {
+        row.cells <- seq(1, all.rows, by = every)
+        col.cells <- seq(1, all.cols, by = every)
+        
+        columns <- c()
+        foreach(rc = row.cells) %do% {
+            columns <- c(columns, c(col.cells + (((rc[1]-1) * all.cols))))
+        }
+        columns <- columns + 1  # Skip "label" column
+            
+        df <- do.call(rbind, alply(df, .margins = 1, function (row) {
+            cell.data <- as.list(row[1, columns])
+            
+            return(data.frame(label = row$label, cell.data))
+        }))
+    }
+    
+    return(df)
+}
+
+if (!exists("source.data")) {
+    max.rows <- 2000
     gunzip("train.csv.gz", overwrite=TRUE, remove = FALSE)
-    train.data <- read.csv("train.csv",
-                           header = TRUE,
-                           nrows=max.rows)
-    train.data$label <- as.factor(train.data$label)
+    source.data <- read.csv("train.csv",
+                            header = TRUE,
+                            nrows=max.rows)
+    source.data$label <- as.factor(source.data$label)
     unlink("train.csv")
 }
+train.data <- source.data
+train.data <- featureEngineering(train.data)
 
 # convert training data into list of labels with a corresponding matrix
 get.list <- function(df = train.data) {
@@ -35,20 +59,6 @@ get.list <- function(df = train.data) {
     return(l)
 }
 
-# train.list <- get.list(train.data)
-# 
-# train.list <- lapply(train.list, function (row) {
-#     .cols = seq(1, cols) * every
-#     .rows = seq(1, rows) * every
-#     
-#     row$figure <- matrix(row$figure[.rows, .cols], rows, cols)
-#     return(row)
-# })
-# 
-# train.data <- ldply(train.list, function (row) {
-#     return(list(label = row$label, unlist(alply(row$figure, 1, function (l) alply(l,1)))))
-# })
-
 figure.columns <- seq(2, (1+rows*cols))
 
 train.data[, figure.columns] <- train.data[, figure.columns] / 255
@@ -61,25 +71,48 @@ start.time <- Sys.time()
 m <- nnet(formula = label ~ .,
           data = train.batch,
           MaxNWts = 100000,
-          maxit = 200,
-          size = (rows*cols/rows))
+          maxit = 400,
+          size = floor(1.5*rows*cols/rows))
 print(Sys.time() - start.time)
 
 test.batch$prediction <- predict(object = m, newdata = test.batch, type="class")
 r <- test.batch[, c("label", "prediction")]
 
 num.correct <- NROW(which(test.batch$label == test.batch$prediction))
-printf("%03.1f%% correct", 100 * num.correct / NROW(test.batch))
+ratio <- num.correct / NROW(test.batch)
 
-# Now run our model against the competition data
-if (!exists("competition.data")) {
-    gunzip("test.csv.gz", overwrite=TRUE, remove = FALSE)
-    competition.data <- read.csv("test.csv", header = TRUE)
-    unlink("test.csv")
+stop()
+
+
+tryCatch(
+    load("bestRatio.RData"),
+    error = function(cond) bestRatio <<- 0
+)
+
+if (ratio > bestRatio) {
+    printf("Best yet: %.3f\n",  ratio)
+    
+    # Now run our model against the competition data
+    if (!exists("competition.data")) {
+        gunzip("test.csv.gz", overwrite=TRUE, remove = FALSE)
+        competition.data <- read.csv("test.csv", header = TRUE)
+        unlink("test.csv")
+    }
+    
+    output <- data.frame(ImageId=1:nrow(competition.data))
+    output$Label <- predict(object = m, newdata = competition.data, type = "class")
+    
+    suffix <- paste(sep="-", format(Sys.time(), "%Y%m%d-%H%M"), round(ratio, 3))
+    fname <- paste0("output/submission-", suffix, ".csv")
+    
+    write.table(output, file=fname, sep = ",", quote = FALSE, row.names = FALSE, col.names = TRUE)
+    gzip(fname, overwrite = TRUE)
+    
+    bestRatio <- ratio
+    save(bestRatio, file="bestRatio.RData")
+    
+    rm(competition.data, output)    # Conserve RAM
+    
+} else {
+    printf("Not an improvement %.3f < %.3f\n", ratio, bestRatio)
 }
-
-predictions <- predict(object = m, newdata = competition.data, type = "class")
-
-output <- data.frame(ImageId=1:nrow(competition.data), Label = predictions)
-write.table(output, file="output.csv", sep = ",", quote = FALSE, row.names = FALSE, col.names = TRUE)
-gzip("output.csv", overwrite = TRUE)
